refactor(e2e): put common setup in setup file
